More tests for decoded JWTs
